+++ playbooks/ansible_connector.py
@@ -217,26 +217,34 @@ def run_install_new_game_server(server_id):
     else:
         run_cmd(pre_install_cmd)
 
-    subcmd1 = sudo_pre_cmd + ["-u", server.username]
-    subcmd2 = [f"{server.install_path}/{server.script_name}", "auto-install"]
-    install_cmd = subcmd1 + subcmd2
+    install_reqs = [f"{server.install_path}/{server.script_name}", "auto-install"]
+
+    # Then run as user to install actual game server.
+    user_prepend = sudo_pre_cmd + ["-u", server.username]
+    install_cmd = user_prepend + install_reqs
 
     if O["dry"]:
         print(install_cmd)
+    else:
+        # Actually run install!
+        run_cmd(install_cmd, server.install_path)
+
+        # Post install cfg fix.
+        post_install_cfg_fix(server.install_path)
+
+    if O["dry"]:
+        print(install_reqs)
         exit()
-
-    # Actually run install!
-    run_cmd(install_cmd, server.install_path)
-
-    # Post install cfg fix.
-    post_install_cfg_fix(server.install_path)
+    else:
+        # After game server install, install required apt reqs as root.
+        run_cmd(install_reqs, server.install_path)
 
     # Cleanup temp sudoers rule.
     tmp_sudoers = f"/etc/sudoers.d/{server.username}-temp-auto-install"
     if os.path.isfile(tmp_sudoers): 
         os.remove(tmp_sudoers)
 
-    # Post install ssh setup.
+    # Post install ssh setup for different users.
     append_new_authorized_key(server)
 
     # Mark finished with new session context.